script_analise/requirements.txt
python-dotenv
pandas
google-generativeai
openai
scikit-learn
scipy